GCP/himas-federated-agents/agent_hospital_a/requirements.txt
# Google ADK
google-adk>=1.13.0,<2.0.0
google-adk[a2a]
google-genai>=1.32.0,<2.0.0

# Google Cloud
google-cloud>=0.34.0,<0.35.0
google-cloud-bigquery>=3.11.0,<4.0.0
google-cloud-aiplatform>=1.111.0,<2.0.0
google-cloud-logging>=3.5.0,<4.0.0
google-cloud-storage>=2.19.0

# ML/Data Processing
tensorflow>=2.15.0
scikit-learn>=1.3.0,<2.0.0
pandas>=2.0.0
db-dtypes>=1.2.0,<2.0.0

# Serialization
protobuf>=5.28.3,<7.0.0

# Security
cryptography>=41.0.0,<42.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0